+++ Tests/BugsnagPerformanceTests/SpanTests.mm
@@ -7,8 +7,8 @@
 //
 
 #import <XCTest/XCTest.h>
-#import "Span.h"
-#import "SpanData.h"
+#import "BugsnagPerformanceSpan+Private.h"
+#import "Utils.h"
 
 using namespace bugsnag;
 
@@ -18,169 +18,233 @@ using namespace bugsnag;
 
 @implementation SpanTests
 
-static std::shared_ptr<Span> spanWithStartTime(CFAbsoluteTime startTime, OnSpanEnd onEnd) {
+static BugsnagPerformanceSpan *spanWithStartTime(CFAbsoluteTime startTime, OnSpanClosed onEnded) {
     TraceId tid = {.value = 1};
-    return std::make_shared<Span>(@"test", tid, 1, 0, startTime, BSGFirstClassUnset, onEnd);
+    return [[BugsnagPerformanceSpan alloc] initWithName:@"test"
+                                                traceId:tid
+                                                 spanId:1
+                                               parentId:0
+                                              startTime:startTime
+                                             firstClass:BSGFirstClassUnset
+                                            onSpanClosed:onEnded];
 }
 
 - (void)testStartEndUnset {
     CFAbsoluteTime startTime = CFABSOLUTETIME_INVALID;
     CFAbsoluteTime endTime = startTime;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, CFAbsoluteTimeGetCurrent(), 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, CFAbsoluteTimeGetCurrent(), 0.001);
 }
 
 - (void)testStartNearPastEndUnset {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 0.0005;
     CFAbsoluteTime endTime = CFABSOLUTETIME_INVALID;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, CFAbsoluteTimeGetCurrent(), 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, CFAbsoluteTimeGetCurrent(), 0.001);
 }
 
 - (void)testStartUnsetEndNearPast {
     CFAbsoluteTime startTime = CFABSOLUTETIME_INVALID;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() - 0.0005;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, endTime, 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, endTime, 0.001);
 }
 
 - (void)testStartUnsetEndNearFuture {
     CFAbsoluteTime startTime = CFABSOLUTETIME_INVALID;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 0.0005;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, endTime, 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, endTime, 0.001);
 }
 
 - (void)testStartNearPastEndNearFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 0.0005;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 0.0005;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, endTime, 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, endTime, 0.001);
 }
 
 - (void)testStartFarPastEndUnset {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 1.0;
     CFAbsoluteTime endTime = CFABSOLUTETIME_INVALID;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqualWithAccuracy(spanData->endTime, CFAbsoluteTimeGetCurrent(), 0.001);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqualWithAccuracy(span.endAbsTime, CFAbsoluteTimeGetCurrent(), 0.001);
 }
 
 - (void)testStartFarPastEndNearPast {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 1.0;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() - 0.0005;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartFarPastEndNearFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 1.0;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 0.0005;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartNowEndFarFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent();
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 1.0;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartFarPastEndFarFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 1.0;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 1.0;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartNearPastEndFarFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() - 0.0001;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 1.0;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartNearFutureEndFarFuture {
     CFAbsoluteTime startTime = CFAbsoluteTimeGetCurrent() + 0.0001;
     CFAbsoluteTime endTime = CFAbsoluteTimeGetCurrent() + 1.0;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testStartEndDistantPast {
     CFAbsoluteTime startTime = 0;
     CFAbsoluteTime endTime = 0;
-    __block std::shared_ptr<SpanData> spanData = nullptr;
-    auto span = spanWithStartTime(startTime, ^(std::shared_ptr<SpanData> data) {spanData = data;});
-    span->end(endTime);
-    XCTAssertEqual(spanData->endTime, endTime);
+    __block BugsnagPerformanceSpan *foundSpan= nil;
+    BugsnagPerformanceSpan *span = spanWithStartTime(startTime, ^(BugsnagPerformanceSpan *cbSpan) {foundSpan = cbSpan;});
+    [span endWithAbsoluteTime:endTime];
+    XCTAssertEqual(span.endAbsTime, endTime);
 }
 
 - (void)testAddRemoveAttributes {
-    auto span = spanWithStartTime(0, ^(std::shared_ptr<SpanData>) {});
+    auto span = spanWithStartTime(0, ^(BugsnagPerformanceSpan *) {});
 
-    XCTAssertNil(span->getAttribute(@"a"));
-    span->setAttribute(@"a", @(1));
-    XCTAssertEqualObjects(@(1), span->getAttribute(@"a"));
+    XCTAssertNil([span getAttribute:@"a"]);
+    [span setAttribute:@"a" withValue:@(1)];
+    XCTAssertEqualObjects(@(1), [span getAttribute:@"a"]);
 
-    XCTAssertNil(span->getAttribute(@"b"));
-    span->setAttribute(@"b", @(2));
-    XCTAssertEqualObjects(@(1), span->getAttribute(@"a"));
-    XCTAssertEqualObjects(@(2), span->getAttribute(@"b"));
+    XCTAssertNil([span getAttribute:@"b"]);
+    [span setAttribute:@"b" withValue:@(2)];
+    XCTAssertEqualObjects(@(1), [span getAttribute:@"a"]);
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"b"]);
 
-    span->setAttribute(@"a", @(2));
-    XCTAssertEqualObjects(@(2), span->getAttribute(@"a"));
-    XCTAssertEqualObjects(@(2), span->getAttribute(@"b"));
+    [span setAttribute:@"a" withValue:@(2)];
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"a"]);
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"b"]);
 
-    span->setAttribute(@"a", nil);
-    XCTAssertNil(span->getAttribute(@"a"));
-    XCTAssertEqualObjects(@(2), span->getAttribute(@"b"));
+    [span setAttribute:@"a" withValue:nil];
+    XCTAssertNil([span getAttribute:@"a"]);
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"b"]);
 
-    span->setAttribute(@"a", @(100));
-    XCTAssertEqualObjects(@(100), span->getAttribute(@"a"));
-    XCTAssertEqualObjects(@(2), span->getAttribute(@"b"));
+    [span setAttribute:@"a" withValue:@(100)];
+    XCTAssertEqualObjects(@(100), [span getAttribute:@"a"]);
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"b"]);
 
-    span->setAttribute(@"a", nil);
-    span->setAttribute(@"b", nil);
-    XCTAssertNil(span->getAttribute(@"a"));
-    XCTAssertNil(span->getAttribute(@"b"));
+    [span setAttribute:@"a" withValue:nil];
+    [span setAttribute:@"b" withValue:nil];
+    XCTAssertNil([span getAttribute:@"a"]);
+    XCTAssertNil([span getAttribute:@"b"]);
+}
+
+- (void)testMutability {
+    auto span = spanWithStartTime(0, ^(BugsnagPerformanceSpan *) {});
+
+    XCTAssertNil([span getAttribute:@"a"]);
+    [span setAttribute:@"a" withValue:@(1)];
+    XCTAssertEqualObjects(@(1), [span getAttribute:@"a"]);
+    [span setAttribute:@"a" withValue:@(2)];
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"a"]);
+
+    [span setMultipleAttributes:@{@"b": @(3), @"c": @(4)}];
+    XCTAssertEqualObjects(@(3), [span getAttribute:@"b"]);
+    XCTAssertEqualObjects(@(4), [span getAttribute:@"c"]);
+
+    NSDate *now = [NSDate date];
+    [span updateName:@"X"];
+    XCTAssertEqualObjects(@"X", span.name);
+    [span updateStartTime:now];
+    XCTAssertEqualObjects(now, span.startTime);
+    [span updateSamplingProbability:0.5];
+    XCTAssertEqual(0.5, span.samplingProbability);
+
+    [span end];
+
+    [span setAttribute:@"a" withValue:@(3)];
+    XCTAssertEqualObjects(@(2), [span getAttribute:@"a"]);
+
+    XCTAssertNil([span getAttribute:@"x"]);
+    [span setAttribute:@"x" withValue:@(2)];
+    XCTAssertNil([span getAttribute:@"x"]);
+
+    [span setMultipleAttributes:@{@"b": @(10), @"c": @(11)}];
+    XCTAssertEqualObjects(@(3), [span getAttribute:@"b"]);
+    XCTAssertEqualObjects(@(4), [span getAttribute:@"c"]);
+
+    [span setMultipleAttributes:@{@"d": @(3), @"e": @(4)}];
+    XCTAssertNil([span getAttribute:@"d"]);
+    XCTAssertNil([span getAttribute:@"e"]);
+
+    NSDate *previousNow = now;
+    now = [NSDate date];
+    [span updateName:@"Y"];
+    XCTAssertEqualObjects(@"X", span.name);
+    [span updateStartTime:now];
+    XCTAssertEqualObjects(previousNow, span.startTime);
+    [span updateSamplingProbability:0.1];
+    XCTAssertEqual(0.5, span.samplingProbability);
+
+    span.isMutable = true;
+
+    [span setAttribute:@"a" withValue:@(3)];
+    XCTAssertEqualObjects(@(3), [span getAttribute:@"a"]);
+
+    span.isMutable = false;
+
+    [span setAttribute:@"a" withValue:@(4)];
+    XCTAssertEqualObjects(@(3), [span getAttribute:@"a"]);
 }
 
 - (void)testMultithreadedAttributesAccess {
-    auto span = spanWithStartTime(0, ^(std::shared_ptr<SpanData>) {});
+    auto span = spanWithStartTime(0, ^(BugsnagPerformanceSpan *) {});
 
     [NSThread detachNewThreadWithBlock:^{
         for (int i = 0; i < 10000000; i++) {
-            span->setAttributes(@{@"a": @(i)});
+            [span setMultipleAttributes:@{@"a": @(i)}];
         }
     }];
 
     for(int i = 0; i < 1000000; i++) {
-        span->hasAttribute(@"a", @1);
+        [span hasAttribute:@"a" withValue:@1];
     }
 }