+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 Changelog
 =========
 
-## TBD
+## 1.8.0 (2024-08-29)
 
 ### Enhancements
 
 * `service.name` can now be set in the configuration.
   [299](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/299)
+  
+* A fixed `samplingProbability` can now be set in the configuration.
+  [300](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/300)
+
+* OnEnd span callbacks are now called on early spans (spans that ended before library start) once the library is started.
+  [298](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/298)
 
 ## 1.7.0 (2024-08-05)